+++ app/ide-desktop/lib/client/src/security.ts
@@ -108,7 +108,8 @@ function preventNavigation() {
  * https://www.electronjs.org/docs/tutorial/security#13-disable-or-limit-creation-of-new-windows. */
 function disableNewWindowsCreation() {
     electron.app.on('web-contents-created', (_event, contents) => {
-        contents.setWindowOpenHandler(({ url }) => {
+        contents.setWindowOpenHandler(details => {
+            const { url } = details
             const parsedUrl = new URL(url)
             if (TRUSTED_EXTERNAL_HOSTS.includes(parsedUrl.host)) {
                 void electron.shell.openExternal(url)