+++ app/ide-desktop/lib/content/esbuild-config.ts
@@ -1,13 +1,11 @@
-/**
- * @file Configuration for the esbuild bundler and build/watch commands.
+/** @file Configuration for the esbuild bundler and build/watch commands.
  *
  * The bundler processes each entry point into a single file, each with no external dependencies and
  * minified. This primarily involves resolving all imports, along with some other transformations
  * (like TypeScript compilation).
  *
  * See the bundlers documentation for more information:
- * https://esbuild.github.io/getting-started/#bundling-for-node.
- */
+ * https://esbuild.github.io/getting-started/#bundling-for-node. */
 
 import * as childProcess from 'node:child_process'
 import * as fs from 'node:fs/promises'
@@ -48,6 +46,7 @@ export interface PassthroughArguments {
     supportsDeepLinks: boolean
 }
 
+/** Mandatory build options. */
 export interface Arguments extends PassthroughArguments {
     /** List of files to be copied from WASM artifacts. */
     wasmArtifacts: string
@@ -59,9 +58,7 @@ export interface Arguments extends PassthroughArguments {
     ensoglAppPath: string
 }
 
-/**
- * Get arguments from the environment.
- */
+/** Get arguments from the environment. */
 export function argumentsFromEnv(passthroughArguments: PassthroughArguments): Arguments {
     const wasmArtifacts = utils.requireEnv('ENSO_BUILD_GUI_WASM_ARTIFACTS')
     const assetsPath = utils.requireEnv('ENSO_BUILD_GUI_ASSETS')
@@ -74,14 +71,13 @@ export function argumentsFromEnv(passthroughArguments: PassthroughArguments): Ar
 // === Git process ===
 // ===================
 
-/**
- * Get output of a git command.
+/** Get output of a git command.
  * @param command - Command line following the `git` program.
- * @returns Output of the command.
- */
+ * @returns Output of the command. */
 function git(command: string): string {
-    // TODO [mwu] Eventually this should be removed, data should be provided by the build script through `BUILD_INFO`.
-    //            The bundler configuration should not invoke git, it is not its responsibility.
+    // TODO [mwu] Eventually this should be removed, data should be provided by the build script
+    //            through `BUILD_INFO`. The bundler configuration should not invoke git,
+    //            it is not its responsibility.
     return childProcess.execSync(`git ${command}`, { encoding: 'utf8' }).trim()
 }
 
@@ -89,9 +85,7 @@ function git(command: string): string {
 // === Bundling ===
 // ================
 
-/**
- * Generate the builder options.
- */
+/** Generate the builder options. */
 export function bundlerOptions(args: Arguments) {
     const {
         outputPath,
@@ -136,10 +130,13 @@ export function bundlerOptions(args: Arguments) {
                 // All other files are ESM because of `"type": "module"` in `package.json`.
                 name: 'pkg-js-is-cjs',
                 setup: build => {
-                    build.onLoad({ filter: /[/\\]pkg.js$/ }, async ({ path }) => ({
-                        contents: await fs.readFile(path),
-                        loader: 'copy',
-                    }))
+                    build.onLoad({ filter: /[/\\]pkg.js$/ }, async info => {
+                        const { path } = info
+                        return {
+                            contents: await fs.readFile(path),
+                            loader: 'copy',
+                        }
+                    })
                 },
             },
             esbuildPluginCopyDirectories(),
@@ -186,16 +183,15 @@ export function bundlerOptions(args: Arguments) {
 
 /** The basic, common settings for the bundler, based on the environment variables.
  *
- * Note that they should be further customized as per the needs of the specific workflow (e.g. watch vs. build).
- */
+ * Note that they should be further customized as per the needs of the specific workflow
+ * (e.g. watch vs. build). */
 export function bundlerOptionsFromEnv(passthroughArguments: PassthroughArguments) {
     return bundlerOptions(argumentsFromEnv(passthroughArguments))
 }
 
-/** ESBuild options for bundling (one-off build) the package.
+/** esbuild options for bundling the package for a one-off build.
  *
- * Relies on the environment variables to be set.
- */
+ * Relies on the environment variables to be set. */
 export function bundleOptions(passthroughArguments: PassthroughArguments) {
     return bundlerOptionsFromEnv(passthroughArguments)
 }