+++ app/ide-desktop/lib/dashboard/src/authentication/src/authentication/providers/session.tsx
@@ -13,11 +13,12 @@ import * as listen from '../listen'
 // === SessionContext ===
 // ======================
 
+/** State contained in a {@link SessionContext}. */
 interface SessionContextType {
     session: results.Option<cognito.UserSession>
 }
 
-/** See {@link AuthContext} for safety details. */
+/** See `AuthContext` for safety details. */
 const SessionContext = react.createContext<SessionContextType>(
     // eslint-disable-next-line no-restricted-syntax
     {} as SessionContextType
@@ -27,7 +28,8 @@ const SessionContext = react.createContext<SessionContextType>(
 // === SessionProvider ===
 // =======================
 
-interface SessionProviderProps {
+/** Props for a {@link SessionProvider}. */
+export interface SessionProviderProps {
     /** URL that the content of the app is served at, by Electron.
      *
      * This **must** be the actual page that the content is served at, otherwise the OAuth flow will
@@ -45,6 +47,7 @@ interface SessionProviderProps {
     children: react.ReactNode
 }
 
+/** A React provider for the session of the authenticated user. */
 export function SessionProvider(props: SessionProviderProps) {
     const { mainPageUrl, children, userSession, registerAuthEventListener } = props
 
@@ -74,6 +77,8 @@ export function SessionProvider(props: SessionProviderProps) {
      * For example, if a user clicks the signout button, this will clear the user's session, which
      * means we want the login screen to render (which is a child of this provider). */
     react.useEffect(() => {
+        /** Handle Cognito authentication events
+         * @throws {error.UnreachableCaseError} Never. */
         const listener: listen.ListenerCallback = event => {
             switch (event) {
                 case listen.AuthEvent.signIn:
@@ -118,6 +123,7 @@ export function SessionProvider(props: SessionProviderProps) {
 // === useSession ===
 // ==================
 
+/** React context hook returning the session of the authenticated user. */
 export function useSession() {
     return react.useContext(SessionContext)
 }