+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/components/directoryCreateForm.tsx
@@ -8,11 +8,17 @@ import * as error from '../../error'
 import * as modalProvider from '../../providers/modal'
 import CreateForm, * as createForm from './createForm'
 
+// ===========================
+// === DirectoryCreateForm ===
+// ===========================
+
+/** Props for a {@link DirectoryCreateForm}. */
 export interface DirectoryCreateFormProps extends createForm.CreateFormPassthroughProps {
     directoryId: backendModule.DirectoryId
     onSuccess: () => void
 }
 
+/** A form to create a directory. */
 function DirectoryCreateForm(props: DirectoryCreateFormProps) {
     const { directoryId, onSuccess, ...passThrough } = props
     const { backend } = backendProvider.useBackend()