+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/components/fileCreateForm.tsx
@@ -8,11 +8,17 @@ import * as error from '../../error'
 import * as modalProvider from '../../providers/modal'
 import CreateForm, * as createForm from './createForm'
 
+// ======================
+// === FileCreateForm ===
+// ======================
+
+/** Props for a {@link FileCreateForm}. */
 export interface FileCreateFormProps extends createForm.CreateFormPassthroughProps {
     directoryId: backendModule.DirectoryId
     onSuccess: () => void
 }
 
+/** A form to create a file. */
 function FileCreateForm(props: FileCreateFormProps) {
     const { directoryId, onSuccess, ...passThrough } = props
     const { backend } = backendProvider.useBackend()