+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/components/projectCreateForm.tsx
@@ -8,12 +8,17 @@ import * as error from '../../error'
 import * as modalProvider from '../../providers/modal'
 import CreateForm, * as createForm from './createForm'
 
+// =========================
+// === ProjectCreateForm ===
+// =========================
+
+/** Props for a {@link ProjectCreateForm}. */
 export interface ProjectCreateFormProps extends createForm.CreateFormPassthroughProps {
     directoryId: backendModule.DirectoryId
     onSuccess: () => void
 }
 
-// FIXME[sb]: Extract shared shape to a common component.
+/** A form to create a project. */
 function ProjectCreateForm(props: ProjectCreateFormProps) {
     const { directoryId, onSuccess, ...passThrough } = props
     const { backend } = backendProvider.useBackend()