+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/dateTime.ts
@@ -1,7 +1,14 @@
 /** @file Utilities for manipulating and displaying dates and times */
-import * as backend from './backend'
 import * as newtype from '../newtype'
 
+// ================
+// === DateTime ===
+// ================
+
+/** A string with date and time, following the RFC3339 specification. */
+export type Rfc3339DateTime = newtype.Newtype<string, 'Rfc3339DateTime'>
+
+/** Formats date time into the preferred format: `YYYY-MM-DD, hh:mm`. */
 export function formatDateTime(date: Date) {
     const year = date.getFullYear()
     const month = date.getMonth().toString().padStart(2, '0')
@@ -11,6 +18,7 @@ export function formatDateTime(date: Date) {
     return `${year}-${month}-${dayOfMonth}, ${hour}:${minute}`
 }
 
+/** Formats a {@link Date} as a {@link Rfc3339DateTime}  */
 export function toRfc3339(date: Date) {
-    return newtype.asNewtype<backend.Rfc3339DateTime>(date.toISOString())
+    return newtype.asNewtype<Rfc3339DateTime>(date.toISOString())
 }