+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/Storage.java
@@ -4,6 +4,7 @@ import org.enso.base.polyglot.Polyglot_Utils;
 import org.enso.table.data.column.builder.object.Builder;
 import org.enso.table.data.column.builder.object.InferredBuilder;
 import org.enso.table.data.column.builder.object.ObjectBuilder;
+import org.enso.table.data.column.operation.CastProblemBuilder;
 import org.enso.table.data.column.operation.map.MapOperationProblemBuilder;
 import org.enso.table.data.column.storage.type.StorageType;
 import org.enso.table.data.mask.OrderMask;
@@ -345,4 +346,6 @@ public abstract class Storage<T> {
     }
     return new LongStorage(data);
   }
+
+  public abstract Storage<?> cast(StorageType targetType, CastProblemBuilder castProblemBuilder);
 }